Typo and other clean up
