+++ realm-annotations/src/main/java/io/realm/annotations/internal/OptionalAPI.java
@@ -7,7 +7,7 @@ import java.lang.annotation.Target;
 
 /**
  * This annotation is used to mark an API as optional with one or more class {@link #dependencies()}. The bytecode
- * transformer will decide on build time to remove the corresponding API if it doesn't full fill the dependencies.
+ * transformer will decide on build time to remove the corresponding API if it doesn't fulfill the dependencies.
  */
 @Retention(RetentionPolicy.CLASS)
 @Target(ElementType.METHOD)
