+++ realm-transformer/src/main/kotlin/io/realm/transformer/util/MutableCollectionExt.kt
@@ -14,26 +14,25 @@
  * limitations under the License.
  */
 
-package io.realm.transformer
+package io.realm.transformer.util
 
 import com.android.SdkConstants
 import com.android.build.api.transform.TransformInput
 import javassist.ClassPool
-import javassist.LoaderClassPath
 import java.io.File
 import java.util.jar.JarFile
 
 // Find all classes and append their names to the given String collection
-fun MutableCollection<TransformInput>.appendAllToClassNames(classNames: MutableCollection<String>) {
-    for (item in this) {
+fun MutableCollection<TransformInput>.appendThisToClassNames(classNames: MutableCollection<String>) {
+    for (item: TransformInput in this) {
         for (input in item.directoryInputs) {
             val dirPath = input.file.absolutePath
             input.file.walkTopDown().filter { it.isFile && it.endsWith(SdkConstants.DOT_CLASS) }.forEach {
-                val className = it.absolutePath.substring(
+                val qualifiedClassName = it.absolutePath.substring(
                         dirPath.length + 1,
                         input.file.absolutePath.length - SdkConstants.DOT_CLASS.length)
                         .replace(File.separatorChar, '.')
-                classNames.add(className)
+                classNames.add(qualifiedClassName)
             }
         }
 
@@ -44,30 +43,19 @@ fun MutableCollection<TransformInput>.appendAllToClassNames(classNames: MutableC
                 // The jar might not using File.separatorChar as the path separator. So we just replace both `\` and
                 // `/`. It depends on how the jar file was created.
                 // See http://stackoverflow.com/questions/13846000/file-separators-of-path-name-of-zipentry
-                val className = path.substring(0, path.length - SdkConstants.DOT_CLASS.length)
+                val qualifiedClassName = path.substring(0, path.length - SdkConstants.DOT_CLASS.length)
                         .replace('/', '.')
                         .replace('\\', '.')
-                classNames.add(className)
+                classNames.add(qualifiedClassName)
             }
         }
     }
 }
 
 // Add all the classes in this collection to the given ClassPool
-fun MutableCollection<TransformInput>.appendAllToClassPool(classPool: ClassPool) {
+fun MutableCollection<TransformInput>.appendThisToClassPool(classPool: ClassPool) {
     this.forEach {
         it.directoryInputs.forEach { classPool.appendClassPath(it.file.absolutePath) }
         it.jarInputs.forEach { classPool.appendClassPath(it.file.absolutePath) }
     }
 }
-
-fun createClassPool() : ClassPool {
-    // Don't use ClassPool.getDefault(). Doing consecutive builds in the same run (e.g. debug+release)
-    // will use a cached object and all the classes will be frozen.
-    val classPool = ClassPool()
-    classPool.appendSystemPath()
-    classPool.appendClassPath(LoaderClassPath(ClassLoader.getSystemClassLoader()))
-
-    return classPool
-}
-
