+++ realm-transformer/src/main/kotlin/io/realm/transformer/util/Utils.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2016 Realm Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.realm.transformer.util
+
+import javassist.ClassPool
+import javassist.LoaderClassPath
+
+fun createClassPool() : ClassPool {
+    // Don't use ClassPool.getDefault(). Doing consecutive builds in the same run (e.g. debug+release)
+    // will use a cached object and all the classes will be frozen.
+    val classPool = ClassPool()
+    classPool.appendSystemPath()
+    classPool.appendClassPath(LoaderClassPath(ClassLoader.getSystemClassLoader()))
+
+    return classPool
+}
+
